+++ setup.cfg
@@ -42,7 +42,7 @@ ignore_errors = True
 [mypy-prefect.tasks.*]
 ignore_errors = True
 
-[mypy-prefect.utilities.google.*]
+[mypy-prefect.utilities.gcp.*]
 ignore_errors = True
 
 [mypy-tests.*]
