+++ src/prefect/tasks/google/bigquery.py
@@ -8,7 +8,7 @@ from google.oauth2.service_account import Credentials
 from prefect.client import Secret
 from prefect.core import Task
 from prefect.engine.signals import SUCCESS
-from prefect.utilities.google import get_bigquery_client
+from prefect.utilities.gcp import get_bigquery_client
 from prefect.utilities.tasks import defaults_from_attrs
