+++ src/prefect/tasks/google/storage.py
@@ -6,7 +6,7 @@ from google.cloud.exceptions import NotFound
 from prefect import context
 from prefect.client import Secret
 from prefect.core import Task
-from prefect.utilities.google import get_storage_client
+from prefect.utilities.gcp import get_storage_client
 from prefect.utilities.tasks import defaults_from_attrs
