+++ tests/utilities/test_gcp.py
@@ -3,7 +3,7 @@ from unittest.mock import MagicMock
 
 pytest.importorskip("google.cloud")
 
-from prefect.utilities.google import (
+from prefect.utilities.gcp import (
     get_google_client,
     get_storage_client,
     get_bigquery_client,
@@ -13,7 +13,7 @@ from prefect.utilities.google import (
 def test_credentials_are_not_required(monkeypatch):
     submodule = MagicMock()
     creds_loader = MagicMock(return_value=MagicMock(project_id="mocked-proj"))
-    monkeypatch.setattr("prefect.utilities.google.Credentials", creds_loader)
+    monkeypatch.setattr("prefect.utilities.gcp.Credentials", creds_loader)
 
     client = get_google_client(submodule)
 
@@ -27,7 +27,7 @@ def test_credentials_are_used(monkeypatch):
     creds_loader = MagicMock()
     creds = MagicMock(project_id="mocked-proj")
     creds_loader.from_service_account_info.return_value = creds
-    monkeypatch.setattr("prefect.utilities.google.Credentials", creds_loader)
+    monkeypatch.setattr("prefect.utilities.gcp.Credentials", creds_loader)
     client = get_google_client(submodule, credentials=dict())
 
     assert creds_loader.from_service_account_info.called
@@ -42,7 +42,7 @@ def test_provided_project_is_prioritized(monkeypatch):
     creds_loader = MagicMock()
     creds = MagicMock(project_id="mocked-proj")
     creds_loader.from_service_account_info.return_value = creds
-    monkeypatch.setattr("prefect.utilities.google.Credentials", creds_loader)
+    monkeypatch.setattr("prefect.utilities.gcp.Credentials", creds_loader)
     client = get_google_client(submodule, credentials=dict(), project="my-proj")
 
     assert creds_loader.from_service_account_info.called
